refactor(releases): extract shared schedule helpers and fix review comments
- Extract parseDate, todayMidnight, daysFromNow, formatShort, getProduct,
  releasePhase, milestoneProgress into shared composable
  (useScheduleHelpers.js) to eliminate ~80 lines of duplication between
  ScheduleView and ScheduleWidget
- Fix getProduct regex inconsistency (add case-insensitive flag)
- formatShort accepts opts.year to handle View (with year) vs Widget
  (without year) difference via a single function
