+++ modules/releases/client/composables/useScheduleHelpers.js
@@ -1,0 +1,67 @@
+export function parseDate(val) {
+  if (!val) return null
+  const d = new Date(val)
+  return isNaN(d.getTime()) ? null : d
+}
+
+export function todayMidnight() {
+  const d = new Date()
+  d.setHours(0, 0, 0, 0)
+  return d
+}
+
+export function daysFromNow(dateStr) {
+  const d = parseDate(dateStr)
+  if (!d) return null
+  const today = todayMidnight()
+  d.setHours(0, 0, 0, 0)
+  return Math.ceil((d.getTime() - today.getTime()) / 86400000)
+}
+
+export function formatShort(dateStr, opts) {
+  const d = parseDate(dateStr)
+  if (!d) return '—'
+  const format = { month: 'short', day: 'numeric' }
+  if (opts && opts.year) format.year = 'numeric'
+  return d.toLocaleDateString('en-US', format)
+}
+
+export function getProduct(release) {
+  if (release.productPagesShortname) return release.productPagesShortname
+  const match = release.id.match(/^([a-z]+)-/i)
+  return match ? match[1] : release.id
+}
+
+export function releasePhase(release) {
+  const ms = release.milestones || {}
+  const phases = [
+    { label: 'Planning', until: ms.planningFreeze },
+    { label: 'Feature Dev', until: ms.featureFreeze },
+    { label: 'Code Complete', until: ms.codeFreeze },
+    { label: 'Release Prep', until: ms.ga },
+    { label: 'Released', until: null }
+  ]
+  const today = todayMidnight()
+  let phaseIndex = 0
+  for (let i = 0; i < 4; i++) {
+    const d = parseDate(phases[i].until)
+    if (d && d.getTime() <= today.getTime()) {
+      phaseIndex = i + 1
+    }
+  }
+  return { phaseIndex, phases }
+}
+
+export function milestoneProgress(currentDate, prevDate) {
+  const curr = parseDate(currentDate)
+  if (!curr) return null
+  const prev = prevDate ? parseDate(prevDate) : null
+  const today = todayMidnight()
+  if (today.getTime() >= curr.getTime()) return 100
+  if (!prev) return null
+  const total = curr.getTime() - prev.getTime()
+  if (total <= 0) return 100
+  const elapsed = today.getTime() - prev.getTime()
+  if (elapsed <= 0) return 0
+  return Math.min(100, Math.round((elapsed / total) * 100))
+}
